+++ internal/entities/analytics/item.go
@@ -24,3 +24,20 @@ type ProposalsByMonth struct {
 	PeriodStarted  common.Time `json:"period_started"`
 	ProposalsCount uint64      `json:"proposals_count"`
 }
+
+type ProposalsCount struct {
+	Succeeded uint32 `json:"succeeded"`
+	Finished  uint32 `json:"finished"`
+}
+
+type VoterWithVp struct {
+	Voter      string  `json:"voter"`
+	VpAvg      float32 `json:"vp_avg"`
+	VotesCount uint32  `json:"votes_count"`
+}
+
+type MutualDao struct {
+	DaoId         string  `json:"dao_id"`
+	VotersCount   uint32  `json:"voters_count"`
+	VotersPercent float32 `json:"voters_percent"`
+}